% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Cosim.R
\name{run_Cosim}
\alias{run_Cosim}
\title{run the co-simulation through BCVTB's command line.}
\usage{
run_Cosim(bcvtbpath, xmlpath, output)
}
\arguments{
\item{bcvtbpath}{The installtion of BCVTB. For me, it is `'D:/bcvtb'`}

\item{xmlpath}{See xmlpath parameters in \code{\link{copy_xml}}}

\item{output}{See output parameters in \code{\link{write_xml}}}
}
\value{
A xml file. Note: xml_copy function returns no value.
}
\description{
Run the co-simulation through BCVTB's command line.
}
